feat: create proto-crude of StockOperation and ItemOperation

+++ src/main/java/com/example/proj_1_web2/productStore/ProductStoreRepository.java
@@ -2,5 +2,5 @@ package com.example.proj_1_web2.productStore;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductStoreRepository extends JpaRepository<ProductStore, Integer> {
+public interface ProductStoreRepository extends JpaRepository<ProductStore, Long> {
 }
